Refactored biome info and feature generation out of terrain generation for better organization
