+++ scripts/terrainGeneration.js
@@ -10,43 +10,11 @@ import { WORLD_HEIGHT, CHUNK_SIZE } from './chunk.js'
 import { BlockId } from "./blockId.js";
 import { blockData, resourceBlockIds } from './blockData.js'
 import { convertWorldPosToChunkIndex } from './world.js';
-import { distanceSquared, lerp } from './utils.js';
+import { biomeStaticData } from './biomeData.js';
+import { Elevation } from './biomeId.js';
 
 // =======================================================================
 // Global variables
-
-export const Elevation = {
-    Ocean: 0,
-    Beach: 1,
-    Land:  2
-};
-
-export const Biome = {
-    // Ocean
-    ColdOcean:        0,
-    Ocean:            1,
-    WarmOcean:        2,
-    // Beach
-    ColdBeach:        3,
-    Beach:            4,
-    WarmBeach:        5,
-    // Land
-    Tundra:           6,
-    Taiga:            7,
-    Desert:           8,
-    Grassland:        9,
-    Savanna:         10,
-    TemperateForest: 11,
-    TropicalForest:  12,
-    Rainforest:      13,
-};
-
-export const BiomeStrings = Object.keys (Biome);
-
-const NORTH = 0;
-const EAST = 1;
-const SOUTH = 2;
-const WEST = 3;
 
 // =======================================================================
 
@@ -84,168 +52,6 @@ export class TerrainGenerator
         // Still based on the main seed so terrain will be reproducable
         const precipitationRNG = new RNG (this.seed + 4321);
         this.precipitationSimplexNoise = new SimplexNoise (precipitationRNG);
-        
-        // Biomes
-        this.biomeSettings = [];
-        // Elevation: Ocean
-        this.biomeSettings.push ({
-            biome: Biome.ColdOcean,
-            elevation: Elevation.Ocean,
-            precipitationLow:    0,
-            precipitationHigh: 500,
-            temperatureLow: -15,
-            temperatureHigh:  0,
-            surfaceBlock: BlockId.Sand,
-            treeGenerator: null,
-            treeDensity: 0.0
-        });
-        this.biomeSettings.push ({
-            biome: Biome.Ocean,
-            elevation: Elevation.Ocean,
-            precipitationLow:    0,
-            precipitationHigh: 500,
-            temperatureLow:   0,
-            temperatureHigh:  15,
-            surfaceBlock: BlockId.Sand,
-            treeGenerator: null,
-            treeDensity: 0.0
-        });
-        this.biomeSettings.push ({
-            biome: Biome.WarmOcean,
-            elevation: Elevation.Ocean,
-            precipitationLow:    0,
-            precipitationHigh: 500,
-            temperatureLow:  15,
-            temperatureHigh: 30,
-            surfaceBlock: BlockId.Sand,
-            treeGenerator: null,
-            treeDensity: 0.0
-        });
-        // Elevation: Beach
-        this.biomeSettings.push ({
-            biome: Biome.ColdBeach,
-            elevation: Elevation.Beach,
-            precipitationLow:    0,
-            precipitationHigh: 500,
-            temperatureLow: -15,
-            temperatureHigh:  0,
-            surfaceBlock: BlockId.Sand,
-            treeGenerator: null,
-            treeDensity: 0.0
-        });
-        this.biomeSettings.push ({
-            biome: Biome.Beach,
-            elevation: Elevation.Beach,
-            precipitationLow:    0,
-            precipitationHigh: 500,
-            temperatureLow:   0,
-            temperatureHigh:  15,
-            surfaceBlock: BlockId.Sand,
-            treeGenerator: null,
-            treeDensity: 0.0
-            
-        });
-        this.biomeSettings.push ({
-            biome: Biome.WarmBeach,
-            elevation: Elevation.Beach,
-            precipitationLow:    0,
-            precipitationHigh: 500,
-            temperatureLow:  15,
-            temperatureHigh: 30,
-            surfaceBlock: BlockId.Sand,
-            treeGenerator: null,
-            treeDensity: 0.0
-        });
-        // Elevation: Land
-        this.biomeSettings.push ({
-            biome: Biome.Tundra,
-            elevation: Elevation.Land,
-            precipitationLow:    0,
-            precipitationHigh: 500,
-            temperatureLow: -15,
-            temperatureHigh: -5,
-            surfaceBlock: BlockId.ColdGrass,
-            treeGenerator: null,
-            treeDensity: 0.0
-        });
-        this.biomeSettings.push ({
-            biome: Biome.Taiga,
-            elevation: Elevation.Land,
-            precipitationLow:   50,
-            precipitationHigh: 500,
-            temperatureLow:  -5,
-            temperatureHigh:  0,
-            surfaceBlock: BlockId.ColdGrass,
-            treeGenerator: this.generatePineTree.bind (this),
-            treeDensity: 0.01
-        });
-        this.biomeSettings.push ({
-            biome: Biome.Desert,
-            elevation: Elevation.Land,
-            precipitationLow:    0,
-            precipitationHigh:  50,
-            temperatureLow:  -5,
-            temperatureHigh: 30,
-            surfaceBlock: BlockId.Sand,
-            treeGenerator: this.generateCactus.bind (this),
-            treeDensity: 0.0025
-        });
-        this.biomeSettings.push ({
-            biome: Biome.Grassland,
-            elevation: Elevation.Land,
-            precipitationLow:   50,
-            precipitationHigh: 150,
-            temperatureLow:   0,
-            temperatureHigh: 15,
-            surfaceBlock: BlockId.Grass,
-            treeGenerator: this.generateOakTree.bind (this),
-            treeDensity: 0.0001
-        });
-        this.biomeSettings.push ({
-            biome: Biome.Savanna,
-            elevation: Elevation.Land,
-            precipitationLow:   50,
-            precipitationHigh: 150,
-            temperatureLow:  15,
-            temperatureHigh: 30,
-            surfaceBlock: BlockId.AridGrass,
-            treeGenerator: this.generateAcaciaTree.bind (this),
-            treeDensity: 0.0005
-        });
-        this.biomeSettings.push ({
-            biome: Biome.TemperateForest,
-            elevation: Elevation.Land,
-            precipitationLow:  150,
-            precipitationHigh: 500,
-            temperatureLow:   0,
-            temperatureHigh: 15,
-            surfaceBlock: BlockId.Grass,
-            treeGenerator: this.generateOakTree.bind (this),
-            treeDensity: 0.009
-        });
-        this.biomeSettings.push ({
-            biome: Biome.TropicalForest,
-            elevation: Elevation.Land,
-            precipitationLow:  150,
-            precipitationHigh: 250,
-            temperatureLow:  15,
-            temperatureHigh: 30,
-            surfaceBlock: BlockId.TropicalGrass,
-            treeGenerator: this.generateOakTree.bind (this),
-            treeDensity: 0.02
-        });
-        this.biomeSettings.push ({
-            biome: Biome.Rainforest,
-            elevation: Elevation.Land,
-            precipitationLow:  250,
-            precipitationHigh: 500,
-            temperatureLow:  15,
-            temperatureHigh: 30,
-            surfaceBlock: BlockId.TropicalGrass,
-            treeGenerator: this.generateJungleTree.bind (this),
-            treeDensity: 0.01
-        });
-
     }
     
     // ===================================================================
@@ -274,7 +80,7 @@ export class TerrainGenerator
 
                 // Place surface block
                 const biome = this.getBiome (x, z);
-                const surfaceBlock = this.biomeSettings[biome].surfaceBlock;
+                const surfaceBlock = biomeStaticData[biome].surfaceBlock;
                 this.world.setBlockId (x, surfaceHeight, z, surfaceBlock);
 
                 // continue;
@@ -364,7 +170,7 @@ export class TerrainGenerator
         const precipitation = this.getPrecipitation (worldX, worldZ);
 
         // Find the biome that matches
-        for (let biomeSettings of this.biomeSettings)
+        for (let biomeSettings of biomeStaticData)
         {
             // Ensure elevation matches
             if (elevation != biomeSettings.elevation)
@@ -502,7 +308,7 @@ export class TerrainGenerator
                 // Determine what biome this is so we know what tree to
                 // spawn and with what frequency
                 const biome = this.getBiome (x, z);
-                const treeDensity = this.biomeSettings[biome].treeDensity;
+                const treeDensity = biomeStaticData[biome].treeDensity;
                 
                 // Ensure that we are going to spawn a tree here
                 const randomValue = vegetationRNG.random ();
@@ -518,7 +324,7 @@ export class TerrainGenerator
                 for ( ; y < WORLD_HEIGHT; ++y)
                 {
                     const blockId = this.world.getBlockId (x, y, z);
-                    if (blockId == this.biomeSettings[biome].surfaceBlock)
+                    if (blockId == biomeStaticData[biome].surfaceBlock)
                         break;
                 }
                 // Ensure that we found the surface
@@ -528,479 +334,11 @@ export class TerrainGenerator
                 const surfaceHeight = y;
 
                 // Spawn tree
-                const treeGenerator
-                    = this.biomeSettings[biome].treeGenerator;
+                const treeGenerator = biomeStaticData[biome].treeGenerator;
                 if (treeGenerator)
-                    treeGenerator (x, surfaceHeight+1, z, vegetationRNG);
+                    treeGenerator (this.world, x, surfaceHeight+1, z, vegetationRNG);
 
             }
-        }
-
-    }
-
-    // ===================================================================
-
-    generateOakTree (x, y, z, vegetationRNG)
-    {
-        // Place trunk
-        const trunkHeightMin = 2;
-        const trunkHeightMax = 6;
-        const trunkHeight = lerp (
-            trunkHeightMin,
-            trunkHeightMax,
-            vegetationRNG.random ()
-        );
-        for (let i = 0; i < trunkHeight; ++i)
-        {
-            this.world.setBlockId (x, y+i, z, BlockId.Log);
-        }
-
-        // Place crown
-        const crownHeightMin = 3;
-        const crownHeightMax = 6;
-        // offsetting coords so rand number is different from above
-        const crownHeight = lerp (
-            crownHeightMin,
-            crownHeightMax,
-            vegetationRNG.random ()
-        );
-        const crownRadiusBottom = 3;
-        const crownRadiusTop = 2;
-        for (let crownRow = 0; crownRow < crownHeight; ++crownRow)
-        {
-            const crownY = y + trunkHeight + crownRow;
-            // Place the part of the trunk that is in the crown
-            if (crownRow < crownHeight - 1)
-                this.world.setBlockId (x, crownY, z, BlockId.Log);
-
-            // Lerp between bottom and top radii while moving up the crown
-            const crownHeightRatio = crownRow / crownHeight;
-            const crownRadius = lerp (
-                crownRadiusBottom,
-                crownRadiusTop,
-                crownHeightRatio
-            );
-            for (let crownX = x-crownRadius+1; crownX < x+crownRadius; ++crownX)
-            {
-                for (let crownZ = z-crownRadius+1; crownZ < z+crownRadius; ++crownZ)
-                {
-                    // Ignore corners
-                    const isCornerX = crownX == x-crownRadius+1
-                        || crownX == x+crownRadius-1;
-                    const isCornerZ = crownZ == z-crownRadius+1
-                        || crownZ == z+crownRadius-1;
-                    if (isCornerX && isCornerZ)
-                        continue;
-                    
-                    // Ensure we are not overwritting blocks
-                    const blockId = this.world.getBlockId (
-                        crownX,
-                        crownY,
-                        crownZ
-                    );
-                    if (blockId != BlockId.Air)
-                        continue;
-
-                    this.world.setBlockId (
-                        crownX,
-                        crownY,
-                        crownZ,
-                        BlockId.Leaves
-                    );
-                }
-            }
-        }
-    }
-
-    // ===================================================================
-
-    generatePineTree (x, y, z, vegetationRNG)
-    {
-        // Place trunk
-        const trunkHeightMin = 2;
-        const trunkHeightMax = 6;
-        const trunkHeight = lerp (
-            trunkHeightMin,
-            trunkHeightMax,
-            vegetationRNG.random ()
-        );
-        for (let i = 0; i < trunkHeight; ++i)
-        {
-            this.world.setBlockId (x, y+i, z, BlockId.PineLog);
-        }
-
-        // Place crown
-        const crownHeightMin = 8;
-        const crownHeightMax = 15;
-        const crownHeight = lerp (
-            crownHeightMin,
-            crownHeightMax,
-            vegetationRNG.random ()
-        );
-        const crownRadiusBottom = crownHeight / 3;
-        const crownRadiusTop = 1;
-        for (let crownRow = 0; crownRow < crownHeight; ++crownRow)
-        {
-            const crownY = y + trunkHeight + crownRow;
-            // Place the part of the trunk that is in the crown
-            if (crownRow < crownHeight - 1)
-                this.world.setBlockId (x, crownY, z, BlockId.PineLog);
-            else
-                this.world.setBlockId (x, crownY, z, BlockId.PineLeaves);
-
-            // Lerp between bottom and top radii while moving up the crown
-            const crownHeightRatio = crownRow / crownHeight;
-            let crownRadius = lerp (
-                crownRadiusBottom,
-                crownRadiusTop,
-                crownHeightRatio
-            );
-            // adjust so that alternating layers push in and out
-            if (crownRow % 2 == 1)
-                crownRadius -= 1;
-            const crownRadiusSquared = crownRadius * crownRadius;
-            for (let crownX = x-crownRadius+1; crownX < x+crownRadius; ++crownX)
-            {
-                for (let crownZ = z-crownRadius+1; crownZ < z+crownRadius; ++crownZ)
-                {
-                    // Ignore blocks outside of radius
-                    const distSquared = distanceSquared (
-                        x,
-                        z,
-                        crownX,
-                        crownZ
-                    );
-                    if (distSquared > crownRadiusSquared)
-                        continue;
-                    
-                    // Ensure we are not overwritting blocks
-                    const blockId = this.world.getBlockId (
-                        crownX,
-                        crownY,
-                        crownZ
-                    );
-                    if (blockId != BlockId.Air)
-                        continue;
-
-                    this.world.setBlockId (
-                        crownX,
-                        crownY,
-                        crownZ,
-                        BlockId.PineLeaves
-                    );
-                }
-            }
-        }
-    }
-
-    // ===================================================================
-
-    generateAcaciaTree (x, y, z, vegetationRNG)
-    {
-        // Place trunk
-        const trunkHeightMin = 1;
-        const trunkHeightMax = 4;
-        const trunkHeight = lerp (
-            trunkHeightMin,
-            trunkHeightMax,
-            vegetationRNG.random ()
-        );
-        for (let i = 0; i < trunkHeight; ++i)
-        {
-            this.world.setBlockId (x, y+i, z, BlockId.AcaciaLog);
-        }
-
-        const dir = lerp (0, 3, vegetationRNG.random ());
-        const height1 = lerp (3, 5, vegetationRNG.random ());
-        this.placeAcaciaBranch (x, y+trunkHeight, z, dir, 0, height1);
-        const oppositeDir = (dir + 2) % 4;
-        const height2 = lerp (1, 3, vegetationRNG.random ());
-        this.placeAcaciaBranch (
-            x,
-            y+trunkHeight,
-            z,
-            oppositeDir,
-            0,
-            height2
-        );
-
-    }
-    
-    // ===================================================================
-
-    placeAcaciaCanopy (x, y, z)
-    {
-        // Place crown
-        const crownHeight = 2
-        const crownRadiusBottom = 4;
-        const crownRadiusTop = 2;
-        for (let crownRow = 0; crownRow < crownHeight; ++crownRow)
-        {
-            const crownY = y + crownRow;
-
-            // Lerp between bottom and top radii while moving up the crown
-            const crownHeightRatio = crownRow / crownHeight;
-            const crownRadius = lerp (
-                crownRadiusBottom,
-                crownRadiusTop,
-                crownHeightRatio
-            );
-            const crownRadiusSquared = crownRadius * crownRadius;
-            for (let crownX = x-crownRadius+1; crownX < x+crownRadius; ++crownX)
-            {
-                for (let crownZ = z-crownRadius+1; crownZ < z+crownRadius; ++crownZ)
-                {
-                    // Ignore blocks outside of radius
-                    const distSquared = distanceSquared (
-                        x,
-                        z,
-                        crownX,
-                        crownZ
-                    );
-                    if (distSquared > crownRadiusSquared)
-                        continue;
-                    
-                    // Ensure we are not overwritting blocks
-                    const blockId = this.world.getBlockId (
-                        crownX,
-                        crownY,
-                        crownZ
-                    );
-                    if (blockId != BlockId.Air)
-                        continue;
-
-                    this.world.setBlockId (
-                        crownX,
-                        crownY,
-                        crownZ,
-                        BlockId.AcaciaLeaves
-                    );
-                }
-            }
-        }
-    }
-
-    // ===================================================================
-
-    placeAcaciaBranch (x, y, z, dir, length, limit)
-    {
-        if (length >= limit)
-        {
-            this.placeAcaciaCanopy (x, y, z);
-            return;
-        }
-        if (dir == NORTH)
-        {
-            this.world.setBlockId (x, y, z, BlockId.AcaciaLog);
-            this.world.setBlockId (x, y, z+1, BlockId.AcaciaLog);
-            this.placeAcaciaBranch (x, y+1, z+1, dir, length+1, limit);
-        }
-        else if (dir == EAST)
-        {
-            this.world.setBlockId (x, y, z, BlockId.AcaciaLog);
-            this.world.setBlockId (x+1, y, z, BlockId.AcaciaLog);
-            this.placeAcaciaBranch (x+1, y+1, z, dir, length+1, limit);
-        }
-        else if (dir == SOUTH)
-        {
-            this.world.setBlockId (x, y, z, BlockId.AcaciaLog);
-            this.world.setBlockId (x, y, z-1, BlockId.AcaciaLog);
-            this.placeAcaciaBranch (x, y+1, z-1, dir, length+1, limit);
-        }
-        else if (dir == WEST)
-        {
-            this.world.setBlockId (x, y, z, BlockId.AcaciaLog);
-            this.world.setBlockId (x-1, y, z, BlockId.AcaciaLog);
-            this.placeAcaciaBranch (x-1, y+1, z, dir, length+1, limit);
-        }
-    }
-
-    // ===================================================================
-
-    generateJungleTree (x, y, z, vegetationRNG)
-    {
-        // Place trunk
-        const trunkHeight = lerp (15, 25, vegetationRNG.random ());
-        for (let i = 0; i < trunkHeight; ++i)
-        {
-            this.world.setBlockId (x, y+i, z, BlockId.JungleLog);
-        }
-        // place tree roots
-        const rootOffset = lerp (0, 3, vegetationRNG.random ());
-        this.placeJungleTrunkRoot (x, y+rootOffset, z, NORTH, 0, 4);
-        this.placeJungleTrunkRoot (x, y+rootOffset, z, EAST , 0, 4);
-        this.placeJungleTrunkRoot (x, y+rootOffset, z, SOUTH, 0, 4);
-        this.placeJungleTrunkRoot (x, y+rootOffset, z, WEST , 0, 4);
-
-        // Side branches
-        const numBranches = lerp (1, 3, vegetationRNG.random ());
-        for (let b = 0; b < numBranches; ++b)
-        {
-            const branchY = y + lerp (
-                4,
-                trunkHeight,
-                vegetationRNG.random ()
-            );
-            const dir = lerp (0, 3, vegetationRNG.random ());
-            const heightLimit = lerp (1, 4, vegetationRNG.random ());
-            this.placeJungleBranch (x, branchY, z, dir, 0, heightLimit);
-        }
-
-        // Top branches
-        const numTopBranches = lerp (0, 2, vegetationRNG.random ());
-        for (let b = 0; b < numTopBranches; ++b)
-        {
-            const dir = lerp (0, 3, vegetationRNG.random ());
-            const heightLimit = lerp (2, 5, vegetationRNG.random ());
-            this.placeJungleBranch (
-                x,
-                y+trunkHeight,
-                z,
-                dir,
-                0,
-                heightLimit
-            );
-        }
-
-        // Put a canopy topper at the top of the trunk
-        this.placeJungleCanopy (x, y+trunkHeight, z);
-
-    }
-
-    // ===================================================================
-
-    placeJungleTrunkRoot (x, y, z, dir, length, limit)
-    {
-        if (length >= limit)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            return;
-        }
-        if (dir == NORTH)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.world.setBlockId (x, y, z+1, BlockId.JungleLog);
-            this.placeJungleTrunkRoot (x, y-1, z+1, dir, length+1, limit);
-        }
-        else if (dir == EAST)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.world.setBlockId (x+1, y, z, BlockId.JungleLog);
-            this.placeJungleTrunkRoot (x+1, y-1, z, dir, length+1, limit);
-        }
-        else if (dir == SOUTH)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.world.setBlockId (x, y, z-1, BlockId.JungleLog);
-            this.placeJungleTrunkRoot (x, y-1, z-1, dir, length+1, limit);
-        }
-        else if (dir == WEST)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.world.setBlockId (x-1, y, z, BlockId.JungleLog);
-            this.placeJungleTrunkRoot (x-1, y-1, z, dir, length+1, limit);
-        }
-    }
-
-    // ===================================================================
-
-    placeJungleCanopy (x, y, z)
-    {
-        // Place crown
-        const crownHeight = 2
-        const crownRadiusBottom = 4;
-        const crownRadiusTop = 2;
-        for (let crownRow = 0; crownRow < crownHeight; ++crownRow)
-        {
-            const crownY = y + crownRow;
-
-            // Lerp between bottom and top radii while moving up the crown
-            const crownHeightRatio = crownRow / crownHeight;
-            const crownRadius = lerp (
-                crownRadiusBottom,
-                crownRadiusTop,
-                crownHeightRatio
-            );
-            const crownRadiusSquared = crownRadius * crownRadius;
-            for (let crownX = x-crownRadius+1; crownX < x+crownRadius; ++crownX)
-            {
-                for (let crownZ = z-crownRadius+1; crownZ < z+crownRadius; ++crownZ)
-                {
-                    // Ignore blocks outside of radius
-                    const distSquared = distanceSquared (
-                        x,
-                        z,
-                        crownX,
-                        crownZ
-                    );
-                    if (distSquared > crownRadiusSquared)
-                        continue;
-                    
-                    // Ensure we are not overwritting blocks
-                    const blockId = this.world.getBlockId (
-                        crownX,
-                        crownY,
-                        crownZ
-                    );
-                    if (blockId != BlockId.Air)
-                        continue;
-
-                    // Place the block
-                    this.world.setBlockId (
-                        crownX,
-                        crownY,
-                        crownZ,
-                        BlockId.JungleLeaves
-                    );
-                }
-            }
-        }
-    }
-
-    // ===================================================================
-
-    placeJungleBranch (x, y, z, dir, length, limit)
-    {
-        if (length >= limit)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.placeJungleCanopy (x, y, z);
-            return;
-        }
-        if (dir == NORTH)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.world.setBlockId (x, y, z+1, BlockId.JungleLog);
-            this.placeJungleBranch (x, y+1, z+1, dir, length+1, limit);
-        }
-        else if (dir == EAST)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.world.setBlockId (x+1, y, z, BlockId.JungleLog);
-            this.placeJungleBranch (x+1, y+1, z, dir, length+1, limit);
-        }
-        else if (dir == SOUTH)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.world.setBlockId (x, y, z-1, BlockId.JungleLog);
-            this.placeJungleBranch (x, y+1, z-1, dir, length+1, limit);
-        }
-        else if (dir == WEST)
-        {
-            this.world.setBlockId (x, y, z, BlockId.JungleLog);
-            this.world.setBlockId (x-1, y, z, BlockId.JungleLog);
-            this.placeJungleBranch (x-1, y+1, z, dir, length+1, limit);
-        }
-    }
-
-    // ===================================================================
-
-    generateCactus (x, y, z, vegetationRNG)
-    {
-        const cactusHeight = lerp (1, 4, vegetationRNG.random ());
-        for (let yoff = 0; yoff < cactusHeight; ++yoff)
-        {
-            this.world.setBlockId (x, y + yoff, z, BlockId.Cactus);
         }
     }
 
@@ -1057,6 +395,5 @@ export class TerrainGenerator
             this.world.addEntity (entity);
         }
     }
-
 
 }
